# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/me": {
      "get": {
        "operationId": "UsersController_me",
        "summary": "Get current user profile (me)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request / Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "patch": {
        "operationId": "UsersController_updateMe",
        "summary": "Update current user profile (me)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated current user profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request / Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/users/me/avatar": {
      "post": {
        "operationId": "UsersController_uploadAvatar",
        "summary": "Upload avatar for current user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Avatar image upload",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "avatar": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "avatar"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated current user profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request / Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpErrorDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Decide API",
    "description": "API documentation",
    "version": "1.0.0",
    "contact": {},
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "https://backend-de-cizhen.onrender.com",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "refreshToken": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken"
      }
    },
    "schemas": {
      "HttpErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 401
          },
          "message": {
            "oneOf": [
              {
                "type": "string",
                "example": "Unauthorized"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "email must be an email"
                ]
              }
            ]
          },
          "error": {
            "type": "string",
            "example": "Unauthorized"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-01-27T10:20:30.123Z"
          },
          "path": {
            "type": "string",
            "example": "/auth/login"
          },
          "requestId": {
            "type": "string",
            "example": "a9f7c8f0-1b2c-4d5e-9f00-123456789abc",
            "description": "Request correlation id (x-request-id)"
          },
          "stack": {
            "type": "string",
            "example": "stack trace...",
            "description": "Only in development"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error",
          "timestamp",
          "path"
        ]
      },
...
      "AvatarDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "example": "/avatars/default.png"
          },
          "isDefault": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "url",
          "isDefault"
        ]
      },
      "MeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "64f0c1a2b3c4d5e6f7a8b9c0"
          },
          "name": {
            "type": "string",
            "example": "Liliia"
          },
          "email": {
            "type": "string",
            "example": "liliia@test.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "client",
              "provider",
              "admin"
            ],
            "example": "client"
          },
          "city": {
            "type": "string",
            "example": "Berlin"
          },
          "language": {
            "type": "string",
            "example": "de"
          },
          "phone": {
            "type": "string",
            "example": "+49123456789"
          },
          "avatar": {
            "$ref": "#/components/schemas/AvatarDto"
          },
          "acceptedPrivacyPolicy": {
            "type": "boolean",
            "example": true
          },
          "acceptedPrivacyPolicyAt": {
            "type": "object",
            "example": "2025-01-01T12:00:00.000Z"
          },
          "isBlocked": {
            "type": "boolean",
            "example": false
          },
          "blockedAt": {
            "type": "object",
            "example": null
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-01-01T12:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-01-02T12:00:00.000Z"
          },
          "lastSeenAt": {
            "type": "object",
            "example": "2026-02-11T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "role",
          "acceptedPrivacyPolicy",
          "isBlocked",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateMeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Liliia"
          },
          "city": {
            "type": "string",
            "example": "Berlin"
          },
          "language": {
            "type": "string",
            "example": "de"
          },
          "phone": {
            "type": "string",
            "example": "+49123456789"
          },
          "avatarUrl": {
            "type": "string",
            "example": "https://cdn.example.com/u/1.png"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "access-token": []
    }
  ]
}